refactor: move query functions into spanner.go
